requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.15
django-extensions==3.2.3
django-restframework==0.0.1
djangorestframework==3.15.2
extensions==0.4
flake8==7.1.1
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
sqlparse==0.5.1
typing-extensions==4.12.2
